requirements.txt
boto3
requests
daily-python
pillow
sounddevice
numpy
